move weight to load::weight
